:+1: [FE] Moved viewmodels to view_models
