+++ frontend/lib/view_models/chat.dart
