+++ frontend/lib/view_models/user.dart
