+++ frontend/lib/views/pages/chat/chat.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:otomo/viewmodels/chat.dart';
+import 'package:otomo/view_models/chat.dart';
 import 'package:otomo/views/bases/indicators/app_circular_progress_indicator.dart';
 import 'package:otomo/views/cases/chat/chat_modal_ui_leading.dart';
 import 'package:otomo/views/cases/chat/chat_ui.dart';
